token no commit + mensagem de error
